Lift navigation definition and parsing to the Chrome layer
