+++ packages/core/chrome/core-chrome-browser-internal/src/project_navigation/index.ts
@@ -7,3 +7,5 @@
  */
 
 export { ProjectNavigationService } from './project_navigation_service';
+
+export type { CloudURLs } from './cloud_links';
