+++ packages/core/chrome/core-chrome-browser-internal/src/project_navigation/navigation_presets.ts
