+++ packages/core/chrome/core-chrome-browser-internal/tsconfig.json
@@ -41,6 +41,10 @@
     "@kbn/core-custom-branding-common",
     "@kbn/core-analytics-browser-mocks",
     "@kbn/core-analytics-browser",
+    "@kbn/default-nav-analytics",
+    "@kbn/default-nav-ml",
+    "@kbn/default-nav-management",
+    "@kbn/default-nav-devtools",
     "@kbn/shared-ux-router",
     "@kbn/shared-ux-link-redirect-app",
     "@kbn/core-http-browser-internal",
