+++ packages/shared-ux/chrome/navigation/tsconfig.json
@@ -20,10 +20,6 @@
     "@kbn/core-chrome-browser",
     "@kbn/core-application-browser",
     "@kbn/i18n",
-    "@kbn/default-nav-analytics",
-    "@kbn/default-nav-ml",
-    "@kbn/default-nav-management",
-    "@kbn/default-nav-devtools",
     "@kbn/shared-ux-storybook-mock",
     "@kbn/core-http-browser",
     "@kbn/cloud-plugin",
